PP-9955 refactor `clients` folder structure
- refactor `clients` folder structure to follow GDS' project directory structure,
  move clients from `services` to be under `app`
