+++ app/clients/products/product.get-by-gateway-account-id.pact.test.js
@@ -5,9 +5,9 @@ const { Pact } = require('@pact-foundation/pact')
 const { expect } = require('chai')
 const proxyquire = require('proxyquire')
 
-const { PactInteractionBuilder } = require('../../../../test/test-helpers/pact/pact-interaction-builder')
-const productFixtures = require('../../../../test/fixtures/product.fixtures')
-const { pactify } = require('../../../../test/test-helpers/pact/pact-base')()
+const { PactInteractionBuilder } = require('../../../test/test-helpers/pact/pact-interaction-builder')
+const productFixtures = require('../../../test/fixtures/product.fixtures')
+const { pactify } = require('../../../test/test-helpers/pact/pact-base')()
 
 // Constants
 const API_RESOURCE = '/v1/api'
@@ -19,7 +19,7 @@ let gatewayAccountId
 
 function getProductsClient (baseUrl = `http://localhost:${port}`, productsApiKey = 'ABC1234567890DEF') {
   return proxyquire('./products.client', {
-    '../../../../config': {
+    '../../../config': {
       PRODUCTS_URL: baseUrl
     }
   })
