+++ app/clients/products/products.client.js
@@ -1,10 +1,10 @@
 'use strict'
 
 // Local Dependencies
-const Product = require('../../../models/Product.class')
-const Payment = require('../../../models/Payment.class')
+const Product = require('../../models/Product.class')
+const Payment = require('../../models/Payment.class')
 const baseClient = require('../base/base.client')
-const { PRODUCTS_URL } = require('../../../../config')
+const { PRODUCTS_URL } = require('../../../config')
 
 // Constants
 const SERVICE_NAME = 'products'
