+++ app/payment-links/confirm/confirm.controller.js
@@ -6,7 +6,7 @@ const paths = require('../../paths')
 const replaceParamsInPath = require('../../utils/replace-params-in-path')
 const captcha = require('../../utils/captcha')
 const logger = require('../../utils/logger')(__filename)
-const productsClient = require('../../services/clients/products/products.client')
+const productsClient = require('../../clients/products/products.client')
 const paymentLinkSession = require('../utils/payment-link-session')
 const { convertPenceToPoundsAndPence } = require('../../utils/currency')
 const { AccountCannotTakePaymentsError } = require('../../errors')
